task 14 changes

+++ 14-binary_tree_balance.c
@@ -9,7 +9,7 @@ int binary_tree_balance(const binary_tree_t *tree)
 {
     size_t left_height = 0;
     size_t right_height = 0;
-    size_t balance = 0;
+    int balance = 0;
 
     if (tree == NULL)
     {
@@ -21,5 +21,5 @@ int binary_tree_balance(const binary_tree_t *tree)
 
     balance = left_height - right_height;
 
-    return ((int) balance);
+    return (balance);
 }
